Funcionando solamente con el component Product
