+++ src/app/components/about/about.component.html
@@ -1,3 +1,17 @@
-<h1 class="animated fadeIn fast">Acerca de Pura Vida</h1>
-<hr />
-<p class="animated fadeIn">Este es un proyecto de la Lic. Paula Daiana Morua ... blablabla</p>
+<div class="jumbotron jumbotron-fluid animated fadeIn fast">
+    <div class="container">
+        <h1 class="display-4">Acerca de Pura Vida</h1>
+        <p class="lead">Este es un proyecto de la Lic. Paula Daiana Morua ... blablabla!</p>
+    </div>
+</div>
+<div class="container">
+    <p class="text-justify">
+        Lorem Ipsum es simplemente el texto de relleno de las imprentas y archivos de texto. Lorem Ipsum ha sido el
+        texto de relleno estándar de las industrias desde el año 1500, cuando un impresor (N. del T. persona que se
+        dedica a la imprenta) desconocido usó una galería de textos y los mezcló de tal manera que logró hacer un libro
+        de textos especimen. No sólo sobrevivió 500 años, sino que tambien ingresó como texto de relleno en documentos
+        electrónicos, quedando esencialmente igual al original. Fue popularizado en los 60s con la creación de las hojas
+        "Letraset", las cuales contenian pasajes de Lorem Ipsum, y más recientemente con software de autoedición, como
+        por ejemplo Aldus PageMaker, el cual incluye versiones de Lorem Ipsum.
+    </p>
+</div>
