+++ src/app/components/home/home.component.html
@@ -4,3 +4,15 @@
         <p class="lead">Mallas a todo ritmo. Precio y calidad!</p>
     </div>
 </div>
+<div class="container">
+    <h2 class="">¿Qué es Lorem Ipsum?</h2>
+    <p class="text-justify">
+        Lorem Ipsum es simplemente el texto de relleno de las imprentas y archivos de texto. Lorem Ipsum ha sido el
+        texto de relleno estándar de las industrias desde el año 1500, cuando un impresor (N. del T. persona que se
+        dedica a la imprenta) desconocido usó una galería de textos y los mezcló de tal manera que logró hacer un libro
+        de textos especimen. No sólo sobrevivió 500 años, sino que tambien ingresó como texto de relleno en documentos
+        electrónicos, quedando esencialmente igual al original. Fue popularizado en los 60s con la creación de las hojas
+        "Letraset", las cuales contenian pasajes de Lorem Ipsum, y más recientemente con software de autoedición, como
+        por ejemplo Aldus PageMaker, el cual incluye versiones de Lorem Ipsum.
+    </p>
+</div>
