+++ src/app/components/navbar/navbar.component.css
@@ -1,0 +1,8 @@
+.logotipo {
+    font-weight: bold;
+}
+
+.logotipo .pura {
+    color: #f94;
+    padding-left: 10px;
+}
