+++ src/app/components/navbar/navbar.component.html
@@ -1,8 +1,8 @@
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
     <div class="container">
-        <a class="navbar-brand" [routerLink]="['home']">
+        <a class="navbar-brand logotipo" [routerLink]="['home']">
             <img src="assets/img/A-64.png" width="30" height="30" class="d-inline-block align-top" alt="" />
-            PuraVida
+            <span class="pura">Pura</span>Vida
         </a>
         <button
             class="navbar-toggler"
@@ -17,7 +17,7 @@
         </button>
 
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <ul class="navbar-nav mr-auto">
+            <ul class="navbar-nav ml-auto">
                 <li class="nav-item" RouterLinkActive="active">
                     <a class="nav-link" [routerLink]="['home']">Inicio<span class="sr-only">(current)</span></a>
                 </li>
@@ -28,30 +28,6 @@
                     <a class="nav-link" [routerLink]="['about']">Acerca de</a>
                 </li>
             </ul>
-            <form class="form-inline my-2 my-lg-0">
-                <input class="form-control mr-sm-2" type="search" placeholder="Buscar producto" aria-label="Search" />
-                <button class="btn btn-outline-light my-2 my-sm-0" type="submit">Buscar</button>
-            </form>
         </div>
     </div>
 </nav>
-<!-- <nav class="navbar navbar-expand-lg navbar-inverse bg-inverse bg-faded">
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <form class="form-inline my-2 my-lg-0">
-            <input
-                (keyup)="buscarProducto(buscarTexto.value)"
-                class="form-control mr-sm-2"
-                type="text"
-                placeholder="Buscar producto"
-                #buscarTexto
-            />
-            <button
-                (click)="buscarProducto(buscarTexto.value)"
-                class="btn btn-outline-primary my-2 my-sm-0"
-                type="button"
-            >
-                Buscar
-            </button>
-        </form>
-    </div>
-</nav> -->
