--- src/app/components/productPage/productPage.component.html
@@ -1,20 +1,0 @@
-<h1 class="animated fadeIn">{{product.nombre}} <small>{{product.aparicion}}</small> </h1>
-<hr>
-
-<div class="row animated fadeIn fast">
-  <div class="col-md-4">
-    <img src="{{product.img}}" class="img-fluid" alt="{{product.nombre}}">
-    <br><br>
-    <a [routerLink]="['/products']" class="btn btn-outline-danger btn-block">Regresar</a>
-  </div>
-
-  <div class="col-md-8">
-    <h3>{{product.nombre}}</h3>
-    <hr>
-    <p>{{product.bio}}</p>
-    <div>
-      <img *ngIf="product.casa=='Marvel'" src="/assets/img/marvel.png" class="logo" alt="">
-      <img *ngIf="product.casa=='DC'"     src="/assets/img/dc.png" class="logo" alt="">
-    </div>
-  </div>
-</div>
