--- src/app/components/productPage/productPage.component.ts
@@ -1,26 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { ProductsService } from '../../services/products.service';
-
-@Component({
-    selector: 'app-productPage',
-    templateUrl: './productPage.component.html'
-})
-export class ProductPageComponent {
-    private product: any = {};
-
-    constructor(
-        private activatedRoute: ActivatedRoute,
-        private _productsService: ProductsService,
-        private router: Router
-    ) {
-        this.activatedRoute.params.subscribe(params => {
-            this.product = this._productsService.getProduct(params['id']);
-            console.log(this.product);
-        });
-    }
-
-    verProducto(idx: number) {
-        this.router.navigate(['/productPage', idx]);
-    }
-}
