--- src/app/components/products/products.component.html
@@ -1,17 +1,0 @@
-<h1>Productos</h1>
-<hr />
-
-<div class="card-columns">
-    <div class="card animated fadeIn fast" *ngFor="let product of products; let i = index">
-        <img src="{{ product.img }}" class="card-img-top" alt="{{ product.nombre }}" />
-        <div class="card-body">
-            <h5 class="card-title">{{ product.nombre }}</h5>
-            <p class="card-text">{{ product.bio }}</p>
-            <p class="card-text">
-                <small class="text-muted">{{ product.aparicion }}</small>
-            </p>
-            <button (click)="verProducto(i)" type="button" class="btn btn-outline-primary btn-block">Ver mas...</button>
-            <!-- <a [routerLink]="['/productPage',i]" class="btn btn-outline-primary btn-block">Ver mas Link...</a> -->
-        </div>
-    </div>
-</div>
