+++ src/app/models/Product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    id: number;
+    nombre: string;
+    bio: string;
+    img: string;
+    aparicion: string;
+    casa: string;
+}
