+++ src/main.ts
@@ -1,8 +1,8 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { environment } from './environments/environment';
 
 import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
 
 if (environment.production) {
     enableProdMode();
